+++ ui/test/resources.ts
@@ -9,7 +9,7 @@ import {
   TemplateType,
   TemplateValueType,
 } from 'src/types'
-import {CellNoteVisibility} from 'src/types/dashboards'
+import {NoteVisibility} from 'src/types/dashboards'
 
 export const role = {
   name: '',
@@ -698,5 +698,5 @@ export const cell: Cell = {
   legend: {},
   inView: true,
   note: 'I am the greatest cell!',
-  noteVisibility: CellNoteVisibility.Default,
+  noteVisibility: NoteVisibility.Default,
 }