+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 1.  [#4287](https://github.com/influxdata/chronograf/pull/4287): Add time selector dropdown to CEO
 1.  [#4311](https://github.com/influxdata/chronograf/pull/4311): Add Flux query editor to the Cell Editor Overlay
 1.  [#4319](https://github.com/influxdata/chronograf/pull/4319): Add Flux query editor to the Data Explorer and use same UI as CEO
+1.  [#4353](https://github.com/influxdata/chronograf/pull/4353): Add visualization options to the Data Explorer
 
 
 ### UI Improvements